Extract settings into context, allow nested editors to render TreeView
