+++ packages/lexical-playground/src/App.js
@@ -7,60 +7,19 @@
  * @flow strict
  */
 
-import type {SettingName} from './appSettings';
-
 import PlaygroundEditorTheme from './themes/PlaygroundEditorTheme';
 import * as React from 'react';
-import {useCallback, useState} from 'react';
-import useSettings from './hooks/useSettings';
-import {DEFAULT_SETTINGS} from './appSettings';
 import Editor from './Editor';
+import Settings from './Settings';
 import TestRecorderPlugin from './plugins/TestRecorderPlugin';
 import TypingPerfPlugin from './plugins/TypingPerfPlugin';
 import {SharedHistoryContext} from './context/SharedHistoryContext';
+import {SettingsContext, useSettings} from './context/SettingsContext';
 import LexicalComposer from '@lexical/react/LexicalComposer';
 
-function setURLParam(param: SettingName, value: null | boolean) {
-  const url = new URL(window.location.href);
-  const params = new URLSearchParams(url.search);
-  if (value !== null) {
-    if (params.has(param)) {
-      params.set(param, String(value));
-    } else {
-      params.append(param, String(value));
-    }
-  } else {
-    if (params.has(param)) {
-      params.delete(param);
-    }
-  }
-  url.search = params.toString();
-  window.history.pushState(null, '', url.toString());
-}
-
 function App(): React$Node {
-  const [settings, setSettings] = useState(DEFAULT_SETTINGS);
-  const setOption = useCallback((setting: SettingName, value: boolean) => {
-    setSettings((options) => ({
-      ...options,
-      [(setting: string)]: value,
-    }));
-    if (DEFAULT_SETTINGS[setting] === value) {
-      setURLParam(setting, null);
-    } else {
-      setURLParam(setting, value);
-    }
-  }, []);
-  const [settingsButton, settingsSwitches] = useSettings(settings, setOption);
-  const {
-    measureTypingPerf,
-    isCharLimit,
-    isCharLimitUtf8,
-    isAutocomplete,
-    isRichText,
-    isCollab,
-    showTreeView,
-  } = settings;
+  const {settings} = useSettings();
+  const {measureTypingPerf} = settings;
 
   return (
     <LexicalComposer namespace="PlaygroundEditor" theme={PlaygroundEditorTheme}>
@@ -69,17 +28,9 @@ function App(): React$Node {
           <img src="logo.svg" alt="Lexical Logo" />
         </header>
         <div className="editor-shell">
-          <Editor
-            isCharLimit={isCharLimit}
-            isCharLimitUtf8={isCharLimitUtf8}
-            isAutocomplete={isAutocomplete}
-            isRichText={isRichText}
-            isCollab={isCollab}
-            showTreeView={showTreeView}
-          />
+          <Editor />
         </div>
-        {settingsSwitches}
-        {settingsButton}
+        <Settings />
         <TestRecorderPlugin />
         {measureTypingPerf && <TypingPerfPlugin />}
       </SharedHistoryContext>
@@ -87,4 +38,10 @@ function App(): React$Node {
   );
 }
 
-export default App;
+export default function PlaygroundApp(): React$Node {
+  return (
+    <SettingsContext>
+      <App />
+    </SettingsContext>
+  );
+}
