+++ packages/lexical-playground/src/Settings.js
@@ -1,0 +1,114 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow strict
+ */
+
+import * as React from 'react';
+import {useMemo, useState} from 'react';
+import Switch from './ui/Switch';
+import {useSettings} from './context/SettingsContext';
+
+export default function Settings(): React$Node {
+  const windowLocation = window.location;
+  const {
+    setOption,
+    settings: {
+      measureTypingPerf,
+      isCollab,
+      isRichText,
+      isCharLimit,
+      isCharLimitUtf8,
+      isAutocomplete,
+      showTreeView,
+      showNestedEditorTreeView,
+    },
+  } = useSettings();
+  const [showSettings, setShowSettings] = useState(false);
+  const [port, isSplitScreen, search] = useMemo(() => {
+    const port = windowLocation.port;
+    const parentWindow = window.parent;
+    const search = windowLocation.search;
+    const isSplitScreen =
+      parentWindow && parentWindow.location.pathname === '/split/';
+    return [port, isSplitScreen, search];
+  }, [windowLocation]);
+
+  return (
+    <>
+      <button
+        id="options-button"
+        className={`editor-dev-button ${showSettings ? 'active' : ''}`}
+        onClick={() => setShowSettings(!showSettings)}
+      />
+      {showSettings ? (
+        <div className="switches">
+          {isRichText && (
+            <Switch
+              onClick={() => {
+                setOption('isCollab', !isCollab);
+                window.location.reload();
+              }}
+              checked={isCollab}
+              text="Collaboration"
+            />
+          )}
+          <Switch
+            onClick={() => {
+              if (isSplitScreen) {
+                window.parent.location.href = `http://localhost:${port}/${search}`;
+              } else {
+                window.location.href = `http://localhost:${port}/split/${search}`;
+              }
+            }}
+            checked={isSplitScreen}
+            text="Split Screen"
+          />
+          <Switch
+            onClick={() => setOption('measureTypingPerf', !measureTypingPerf)}
+            checked={measureTypingPerf}
+            text="Measure Perf"
+          />
+          <Switch
+            onClick={() => setOption('showTreeView', !showTreeView)}
+            checked={showTreeView}
+            text="Tree View"
+          />
+          <Switch
+            onClick={() =>
+              setOption('showNestedEditorTreeView', !showNestedEditorTreeView)
+            }
+            checked={showNestedEditorTreeView}
+            text="Nested Editors Tree View"
+          />
+          <Switch
+            onClick={() => {
+              setOption('isRichText', !isRichText);
+              setOption('isCollab', false);
+            }}
+            checked={isRichText}
+            text="Rich Text"
+          />
+          <Switch
+            onClick={() => setOption('isCharLimit', !isCharLimit)}
+            checked={isCharLimit}
+            text="Char Limit"
+          />
+          <Switch
+            onClick={() => setOption('isCharLimitUtf8', !isCharLimitUtf8)}
+            checked={isCharLimitUtf8}
+            text="Char Limit (UTF-8)"
+          />
+          <Switch
+            onClick={() => setOption('isAutocomplete', !isAutocomplete)}
+            checked={isAutocomplete}
+            text="Autocomplete"
+          />
+        </div>
+      ) : null}
+    </>
+  );
+}
