+++ packages/lexical-playground/src/context/SettingsContext.js
@@ -1,0 +1,66 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow strict
+ */
+
+import type {SettingName} from '../appSettings';
+import {DEFAULT_SETTINGS} from '../appSettings';
+
+import * as React from 'react';
+import {createContext, useContext, useMemo, useState, useCallback} from 'react';
+
+type SettingsContextShape = {
+  settings: {[SettingName]: boolean},
+  setOption: (name: SettingName, value: boolean) => void,
+};
+
+const Context: React$Context<SettingsContextShape> = createContext({});
+
+export const SettingsContext = ({
+  children,
+}: {
+  children: React$Node,
+}): React$Node => {
+  const [settings, setSettings] = useState(DEFAULT_SETTINGS);
+  const setOption = useCallback((setting: SettingName, value: boolean) => {
+    setSettings((options) => ({
+      ...options,
+      [(setting: string)]: value,
+    }));
+    if (DEFAULT_SETTINGS[setting] === value) {
+      setURLParam(setting, null);
+    } else {
+      setURLParam(setting, value);
+    }
+  }, []);
+  const contextValue = useMemo(() => {
+    return {setOption, settings};
+  }, [setOption, settings]);
+  return <Context.Provider value={contextValue}>{children}</Context.Provider>;
+};
+
+export const useSettings = (): SettingsContextShape => {
+  return useContext(Context);
+};
+
+function setURLParam(param: SettingName, value: null | boolean) {
+  const url = new URL(window.location.href);
+  const params = new URLSearchParams(url.search);
+  if (value !== null) {
+    if (params.has(param)) {
+      params.set(param, String(value));
+    } else {
+      params.append(param, String(value));
+    }
+  } else {
+    if (params.has(param)) {
+      params.delete(param);
+    }
+  }
+  url.search = params.toString();
+  window.history.pushState(null, '', url.toString());
+}
