+++ packages/lexical-playground/src/index.css
@@ -86,7 +86,6 @@ pre {
   white-space: pre-wrap;
   margin: 1px auto 10px auto;
   max-height: 250px;
-  width: 940px;
   position: relative;
   border-bottom-left-radius: 10px;
   border-bottom-right-radius: 10px;
@@ -715,6 +714,11 @@ i.export {
   user-select: none;
 }
 
+.editor-shell .editor-image .image-caption-container .tree-view-output {
+  margin: 0;
+  border-radius: 0;
+}
+
 .editor-shell .editor-image .image-caption-container {
   display: block;
   position: absolute;
@@ -722,7 +726,6 @@ i.export {
   left: 0;
   right: 0;
   padding: 0;
-  height: 40px;
   margin: 0;
   border-top: 1px solid #fff;
   background-color: rgba(255,255,255,0.9);
@@ -957,7 +960,6 @@ i.chevron-down {
   overflow: hidden;
   padding: 0 0 10px 0;
   margin: auto;
-  width: 940px;
   display: flex;
 }
