+++ packages/lexical-playground/src/nodes/ImageNode.js
@@ -46,6 +46,8 @@ import TableCellActionMenuPlugin from '../plugins/TableActionMenuPlugin';
 import ImagesPlugin from '../plugins/ImagesPlugin';
 import LinkPlugin from '@lexical/react/LexicalLinkPlugin';
 import stylex from 'stylex';
+import TreeViewPlugin from '../plugins/TreeViewPlugin';
+import {useSettings} from '../context/SettingsContext';
 
 const styles = stylex.create({
   contentEditable: {
@@ -393,6 +395,9 @@ function ImageComponent({
   }, [editor]);
 
   const {historyState} = useSharedHistoryContext();
+  const {
+    settings: {showNestedEditorTreeView},
+  } = useSettings();
 
   return (
     <Suspense fallback={null}>
@@ -445,6 +450,7 @@ function ImageComponent({
                 }
                 skipInit={isCollab}
               />
+              {showNestedEditorTreeView && <TreeViewPlugin />}
             </LexicalNestedComposer>
           </div>
         )}
